requirements.txt
fastapi
uvicorn[standard]
jinja2
requests
python-dotenv
pydantic
google-auth
google-auth-oauthlib
google-api-python-client
google-auth-httplib2